backend/requirements.txt
numpy
pandas
fastapi
uvicorn
python-jose
requests
kafka-python